[Refactor] DTO & 모델 분리
